+++ app/routes/_layout.charts.tsx
@@ -1,0 +1,8 @@
+export default function Charts() {
+    return (
+      <div>
+        Charts
+      </div>
+    );
+  }
+
