+++ app/routes/_layout.planning.tsx
@@ -1,0 +1,8 @@
+export default function Planning() {
+    return (
+      <div>
+        Planning
+      </div>
+    );
+  }
+
