+++ app/routes/_layout.transactions.tsx
@@ -1,0 +1,8 @@
+export default function Transactions() {
+    return (
+      <div>
+        Transactions
+      </div>
+    );
+  }
+
